v0.20.0 採購管理-大氣採購 -修正 新增出現錯誤 -新增 列表 大氣應付未付 -新增 列表 運費應付未付 -修正 總金額四捨五入到整數 -新增 列表欄位 備註 -修正 新訂單選擇 廠商、產品 要取得上次單價
銷售管理
-瓦斯桶單價抓最後一次的價格
-結存瓶異常
-檢驗結存瓶拿掉 進場檢驗瓶時入結存瓶
-查詢時依照選擇的進場 出場 來顯示列表

+++ Models/accounting_entry.vb
@@ -16,8 +16,8 @@ Partial Public Class accounting_entry
     Public Property ae_TransactionId As Integer
     Public Property ae_Date As Date
     Public Property ae_s_Id As Integer
-    Public Property ae_Debit As Nullable(Of Single)
-    Public Property ae_Credit As Nullable(Of Single)
+    Public Property ae_Debit As Nullable(Of Double)
+    Public Property ae_Credit As Nullable(Of Double)
 
     Public Overridable Property subject As subject
